NPE if form submitted more than once, fixes #460

+++ dspace-xmlui/src/main/java/org/dspace/app/xmlui/aspect/submission/submit/RemoveSubmissionsAction.java
@@ -56,7 +56,9 @@ public class RemoveSubmissionsAction extends AbstractAction
         	{
         		// If they selected to remove the item then delete everything.
     			WorkspaceItem workspaceItem = WorkspaceItem.find(context, Integer.valueOf(workspaceID));
-    			workspaceItem.deleteAll();
+				if(workspaceItem != null) {
+					workspaceItem.deleteAll();
+				}
         	}
         	context.commit();
     	}
